Claude/add firmware version reporting l fv p4 (#50)
* feat: add firmware version reporting through full stack

- Create src/version.h with BRAMBLE_FIRMWARE_VERSION constant using
  uint32_t encoding: (major << 24) | (minor << 16) | build
- Change firmware_version from uint16_t to uint32_t across all firmware
  structs (RegistrationPayload, NodeInfo, NodeConfiguration,
  RegistryNodeEntry) and function signatures
- Add firmware_version to hub LIST_NODES UART output for Pi consumption
- Update Python API models and endpoints to parse and expose
  firmware_version as formatted string (e.g. "1.0.0")
- Add firmware version display to dashboard NodeDetail and NodeCard
- Bump hub registry version to 2 (breaking change in struct layout)
- Replace hardcoded 0x0100 with BRAMBLE_FIRMWARE_VERSION in main.cpp
  and irrigation_mode.cpp
- Log firmware version at boot

* chore: set initial firmware version to 0.1.0

* feat: read firmware version from VERSION file via CMake

- Add VERSION file at project root (single source of truth: "0.1.0")
- CMake reads VERSION, parses major.minor.build, passes as compile defs
- version.h uses CMake-provided defines with #ifndef fallback for IDEs
- GitHub Actions can bump VERSION before build to set build number
- Also wires version into pico_set_program_version for UF2 metadata

* feat: auto-increment build number on each cmake configure

VERSION file now contains only major.minor (human-edited).
Build number is auto-incremented by CMake on each configure and
stored in build/build_number.txt (not checked in). Clean build
resets to 0; reconfigure bumps it.

GitHub Actions can override with -DBRAMBLE_VERSION_BUILD=N if needed.

* feat: auto-increment build number on each cmake configure

VERSION file now contains major.minor.build (e.g. 0.1.0). CMake reads
it, increments the build number, and writes it back. The file is
versioned so CI picks up where it left off. CI can also override with
-DBRAMBLE_VERSION_BUILD=N if preferred (e.g. using run_number).

---------

+++ src/lora/address_manager.h
@@ -14,7 +14,7 @@ struct NodeInfo {
     uint16_t assigned_address;      // Assigned network address
     uint8_t node_type;              // Node type (sensor/actuator/hybrid)
     uint8_t capabilities;           // Node capabilities flags
-    uint16_t firmware_version;      // Firmware version
+    uint32_t firmware_version;      // Firmware version
     char device_name[16];           // Human readable name
     uint32_t last_seen_time;        // Last communication time (ms since boot)
     uint32_t last_check_time;       // Last time we checked/updated inactive duration
@@ -40,7 +40,7 @@ public:
      * @return Assigned address (0x0000 if registration failed)
      */
     uint16_t registerNode(uint64_t device_id, uint8_t node_type, uint8_t capabilities,
-                          uint16_t firmware_version, const char *device_name);
+                          uint32_t firmware_version, const char *device_name);
 
     /**
      * @brief Check if a device is already registered
